mvpp2x: removed warnings from compilation
Description:
------------
Removed warnings from all compilation options

Reviewed-on: http://vgitil04.il.marvell.com:8080/27104
Tested-by: Yuval Caduri <cyuval@marvell.com>

+++ mvpp2_debug.c
@@ -220,12 +220,12 @@ static void mvPp2RxQueueDetailedShow(struct mvpp2 *priv, struct mvpp2_rx_queue *
 			   i, rx_desc+i, rx_desc[i].status, rx_desc[i].data_size);
 		if (priv->pp2_version == PPV21) {
 			printk("buf_addr=%lx, buf_cookie=%p",
-				   mvpp21_rxdesc_phys_addr_get(rx_desc),
+				   (long unsigned int)mvpp21_rxdesc_phys_addr_get(rx_desc),
 				   mvpp21_rxdesc_cookie_get(rx_desc));
 		}
 		else {
 			printk("buf_addr=%lx, buf_cookie=%p",
-				   mvpp22_rxdesc_phys_addr_get(rx_desc),
+				   (long unsigned int)mvpp22_rxdesc_phys_addr_get(rx_desc),
 				   mvpp22_rxdesc_cookie_get(rx_desc));
 		}
 
@@ -262,9 +262,9 @@ void mvPp2RxqShow(struct mvpp2 *priv, int port, int rxq, int mode)
 	printk("size=%d, pkts_coal=%d, time_coal=%d\n",pp_rxq->size, pp_rxq->pkts_coal, pp_rxq->time_coal);
 
 	printk("first_virt_addr=%p, first_dma_addr=%lx, next_rx_desc=%d, rxq_cccupied=%d, rxq_nonoccupied=%d\n",
-	pp_rxq->first_desc, MVPP2_DESCQ_MEM_ALIGN(pp_rxq->descs_phys), pp_rxq->next_desc_to_proc,
+	pp_rxq->first_desc, (long unsigned int)MVPP2_DESCQ_MEM_ALIGN(pp_rxq->descs_phys), pp_rxq->next_desc_to_proc,
 	mvpp2_rxq_received(pp_port, pp_rxq->id), mvpp2_rxq_free(pp_port, pp_rxq->id));
-	printk("virt_mem_area_addr=%p, dma_mem_area_addr=%lx\n", pp_rxq->desc_mem, pp_rxq->descs_phys);
+	printk("virt_mem_area_addr=%p, dma_mem_area_addr=%lx\n", pp_rxq->desc_mem, (long unsigned int)pp_rxq->descs_phys);
 
 	if (mode)
 		mvPp2RxQueueDetailedShow(priv, pp_rxq);
@@ -428,8 +428,8 @@ void mvPp2TxqShow(struct mvpp2 *priv, int port, int txq, int mode)
 	printk("physical_txq=%d, size=%d, pkts_coal=%d \n",pp_txq->id, pp_txq->size, pp_txq->pkts_coal);
 
 	printk("first_virt_addr=%p, first_dma_addr=%lx, next_tx_desc=%d\n",
-		   pp_txq->first_desc, MVPP2_DESCQ_MEM_ALIGN(pp_txq->descs_phys), pp_txq->next_desc_to_proc);
-	printk("virt_mem_area_addr=%p, dma_mem_area_addr=%lx\n", pp_txq->desc_mem, pp_txq->descs_phys);
+		   pp_txq->first_desc, (long unsigned int)MVPP2_DESCQ_MEM_ALIGN(pp_txq->descs_phys), pp_txq->next_desc_to_proc);
+	printk("virt_mem_area_addr=%p, dma_mem_area_addr=%lx\n", pp_txq->desc_mem, (long unsigned int)pp_txq->descs_phys);
 
 	for_each_online_cpu(cpu) {
 		txq_pcpu = per_cpu_ptr(pp_txq->pcpu, cpu);
@@ -1314,7 +1314,7 @@ void mvpp2_skb_dump(struct sk_buff *skb, int size, int access)
 	void *addr = skb->head + NET_SKB_PAD;
 	uintptr_t memAddr = (uintptr_t)addr;
 
-	printk("skb=%p, buf=%p, ksize=%d\n", skb, skb->head, ksize(skb->head));
+	printk("skb=%p, buf=%p, ksize=%d\n", skb, skb->head, (int)ksize(skb->head));
 
 	if (access == 0)
 		access = 1;
